backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
python-dotenv==1.0.0
feedparser==6.0.11
playwright==1.40.0
